requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
